AbInitioSoftwareBase module

Public interfaces

AbInitioSoftwareBase.load — Function

load(file)

Load data from file to a Dict.

By now, YAML, JSON, and TOML formats are supported. The format is recognized by file extension.

AbInitioSoftwareBase.loads — Function

loads(str, format)

Load data from str to a Dict. Allowed formats are "yaml", "yml", "json" and "toml".

AbInitioSoftwareBase.save — Function

save(file, data)

Save data to file.

By now, YAML, JSON, and TOML formats are supported. The format is recognized by file extension.

If data is a Dict, its keys should be Strings so that load can return the same data.

Warning

Allowed data types can be referenced in JSON.jl documentation and YAML.jl documentation. For TOML format, only AbstractDict type is allowed.

AbInitioSoftwareBase.of_format — Function

of_format(to, from)

Convert from to the format of to. Similar to oftype.