AbInitioSoftwareBase.Inputs.Input — Type

An abstract type representing an input object of ab initio software.

AbInitioSoftwareBase.Inputs.inputstring — Function

Return a String that represents the input of the software. Need to be implemented.

AbInitioSoftwareBase.Inputs.writeinput — Function

writeinput(io::IO, object::Input)
writeinput(file, object::Input, dry_run::Bool = false)

Write an Input object to file or io. Use dry_run = true to preview changes.